# api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Debuglet dispatcher HTTP API"
  version: "1.2"
  description: "Wire contract of the dispatcher's public HTTP API. Bandwidth values are bits per second, timeouts are milliseconds, TESLA delays are seconds, TESLA anchor timestamps are Unix nanoseconds and every other timestamp is Unix seconds unless a field says otherwise. A client states the contract version it requires with the Debuglet-API-Version request header, declared on every operation below; every response carries the version this dispatcher implements in the same header. Requests are authenticated with a server-issued session: a native client presents it as an Authorization bearer token, a browser client in the session_token cookie plus the X-Debuglet-CSRF header on state-changing requests. Each operation states its own 401 and 403. Access to another account's run or payment order answers 404, so a response is no existence oracle. A dispatcher explicitly configured for the local development profile additionally serves the unauthenticated routes below as its own local operator. See docs/API.md for the access matrix and the compatibility and deprecation policy."
servers:
  - url: "/"
    description: "Dispatcher served at the root of its origin."
  - url: "/api"
    description: "Dispatcher served behind an explicit /api path prefix. Every route below is relative to the selected server URL."
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /executors:
    get:
      operationId: "listExecutors"
      summary: "List the registered executors."
      parameters:
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      security: []
      responses:
        "200":
          description: "The registered executors. The body is null when no executor is registered."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutorList"
        "400":
          description: "The Debuglet-API-Version request header requires a contract version this dispatcher does not serve (unsupported_api_version). This route has no other documented 400."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /executors/by-ip:
    get:
      operationId: "getExecutorByIP"
      summary: "Identify the executor registered for a source IP."
      parameters:
        - name: "ip"
          in: "query"
          required: true
          description: "Source IP of the executor."
          schema:
            type: string
        - name: "n"
          in: "query"
          required: false
          description: "Maximum number of recent debuglet IDs. Defaults to 10; must be a positive integer when present."
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      responses:
        "200":
          description: "The executor serving that address, and the requesting account's own recent debuglet IDs on it. Runs of other accounts are never listed."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutorByIPResponse"
        "400":
          description: "The ip parameter is missing, or n is not a positive integer."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: "The request presented no credential, or one that is unknown, expired or revoked (unauthorized). A dispatcher serving the local development profile also serves this route without a credential."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "No executor is registered for that IP."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /executors/{id}/tesla:
    get:
      operationId: "getExecutorTesla"
      summary: "Read an executor's TESLA key schedule parameters."
      parameters:
        - name: "id"
          in: "path"
          required: true
          description: "Executor ID."
          schema:
            type: string
        - name: "Debuglet-API-Version"
          in: "header"
          required: false
          description: "Contract version the client requires, as major or major.minor. A version this dispatcher does not serve is answered 400 with code unsupported_api_version; an absent or empty value states no requirement. The discovery routes /version and /openapi.yaml and the health routes answer whatever this header says."
          schema:
            type: string
      security: []
      responses:
        "200":
          description: "The parameters an external verifier needs to reconstruct chain keys."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutorTeslaResponse"
        "400":
          description: "The Debuglet-API-Version request header requires a contract version this dispatcher does not serve (unsupported_api_version). This route has no other documented 400."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: "No executor with that ID is registered."
          headers:
            Debuglet-API-Version:
              description: "Contract version this dispatcher implements."
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      description: "A session token from /auth/login, presented as \"Authorization: Bearer <token>\". This is what a native client uses; it is not ambient authority, so it needs no CSRF token."
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "session_token"
      description: "The same session token in a cookie, for a browser client. A state-changing request authenticated this way must repeat the session's CSRF token, returned as csrf_token and set in the session_csrf cookie, in the X-Debuglet-CSRF header."
  schemas:
    Error:
      type: object
      description: "Error envelope of every failure. Branch on code; message is a bounded diagnostic for a human and must not be parsed. A minor version may add codes, so treat an unknown code as a plain failure of the status it arrived with."
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: "Stable machine-readable failure code."
          enum:
            - "invalid_request"
            - "invalid_policy"
            - "unknown_executor"
            - "intent_mismatch"
            - "payment_incomplete"
            - "unsupported_payment_method"
            - "payments_disabled"
            - "unsupported_api_version"
            - "unauthorized"
            - "forbidden"
            - "not_found"
            - "capacity_exhausted"
            - "cancel_refused"
            - "method_not_allowed"
            - "unsupported_media_type"
            - "payload_too_large"
            - "internal_error"
            - "service_unavailable"
        message:
          type: string
          description: "Bounded human-readable diagnostic. Never carries a credential, a stack trace or internal database or transport detail."
...
    ExecutorResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - ready
        - last_seen
        - version
        - tesla_delay_sec
        - tesla_anchor_timestamp_ns
        - tesla_anchor_key
        - price_per_bw
        - currency
      properties:
        id:
          type: string
        ready:
          type: boolean
        last_seen:
          type: integer
          format: int64
          description: "Unix seconds of the last observed contact."
        version:
          type: string
          description: "Version string announced by the executor."
        tesla_delay_sec:
          type: integer
          format: int64
          description: "TESLA epoch duration in seconds."
        tesla_anchor_timestamp_ns:
          type: integer
          format: int64
          description: "Unix nanoseconds of TESLA epoch 0."
        tesla_anchor_key:
          type: string
          format: byte
          nullable: true
          description: "Base64-encoded public chain anchor k_0. Null when the executor announced none."
        price_per_bw:
          type: integer
          format: int64
          description: "Price per bit per second and second of run time, in the executor's currency."
        currency:
          type: string
    ExecutorList:
      type: array
      nullable: true
      description: "Null when no executor is registered."
      items:
        $ref: "#/components/schemas/ExecutorResponse"
    ExecutorByIPResponse:
      type: object
      additionalProperties: false
      required:
        - executor_id
        - debuglet_ids
      properties:
        executor_id:
          type: string
        debuglet_ids:
          type: array
          description: "Most recent debuglet IDs, newest first. An executor with no dispatch history reports an empty array."
          items:
            type: string
            format: uuid
    ExecutorTeslaResponse:
      type: object
      additionalProperties: false
      required:
        - executor_id
        - anchor_key
        - anchor_timestamp_ns
        - delay_sec
        - disclosed_epoch
        - disclosed_key
      properties:
        executor_id:
          type: string
        anchor_key:
          type: string
          format: byte
          description: "Base64-encoded k_0. Empty when the executor announced none."
        anchor_timestamp_ns:
          type: integer
          format: int64
          description: "Unix nanoseconds of epoch 0."
        delay_sec:
          type: integer
          format: int64
          description: "Epoch duration in seconds."
        disclosed_epoch:
          type: integer
          format: int64
          description: "Index of the latest disclosed key, 0 when none was disclosed."
        disclosed_key:
          type: string
          format: byte
          description: "Base64-encoded latest disclosed key. Empty when none was disclosed."
...